06-langgraph/pyproject.toml
[project]
name = "henry-langgraph-module"
version = "0.1.0"
description = "Módulo 06 de LangGraph del curso de AI Engineering de Henry"
requires-python = ">=3.12"
dependencies = [
  "langchain>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.3.0",
  "langchain-community>=0.3.0",
  "langgraph>=0.4.0",
  "langgraph-checkpoint>=4.0.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "ipykernel>=6.29.5",
  "jupyterlab>=4.2.0",
  "nbformat>=5.10.4",
  "nbclient>=0.10.2",
  "nbconvert>=7.16.0",
  "rich>=13.7.1",
  "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-mock>=3.12.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.8.0",
  "pytest-asyncio>=0.23.0",
]

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W"]
extend-ignore = ["E501"]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
extraPaths = ["src"]
reportMissingImports = "none"
reportMissingModuleSource = "none"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
